+++ tools/reset_guild_commands.py
@@ -10,9 +10,10 @@ load_dotenv()  # .env読み込み
 
 TOKEN = os.getenv("TOKEN")
 DEV_GUILD = int(os.getenv("DEV_GUILD"))
+PREFIX = os.getenv("PREFIX")
 
 intents = discord.Intents.all()
-bot = commands.Bot(command_prefix="!", intents=intents, help_command=None)
+bot = commands.Bot(command_prefix=PREFIX, intents=intents, help_command=None)
 
 
 @bot.event
@@ -21,8 +22,8 @@ async def on_ready():
     bot.tree.clear_commands(guild=guild)
     await bot.tree.sync(guild=guild)
     print(f"[Akane] Commands for {DEV_GUILD} cleared")
-    game = discord.Game("メンテナンス中 | Under maintenance")
-    await bot.change_presence(status=discord.Status.idle, activity=game)
+    activity = discord.CustomActivity(name="メンテナンス中 | Under maintenance")
+    await bot.change_presence(status=discord.Status.idle, activity=activity)
 
 
 bot.run(os.getenv("TOKEN"))
